test(MSW): move send-code resolver from authHandlers to its own file and wirt it with MSW
